feat(cli): extract CLI package with app skeleton and convert command
Create src/speculators/cli/ package as the new home for CLI wiring,
per RFC #906. Move the convert command from __main__.py into
cli/convert.py, set up the typer app with Pipeline/Tools help panels,
and simplify __main__.py to a thin delegation layer.

Part of vllm-project/speculators#927 (Sub-PR 1 of 6)
